+++ src/pages/task/taskTabs/buildBaronAndAnnotations/AnnotationTicketsTable/AnnotationTicketRow.tsx
@@ -26,7 +26,6 @@ export const AnnotationTicketRow: React.VFC<AnnotationTicketRowProps> = ({
   confidenceScore,
   loading = false,
 }) => {
-  console.log({ confidenceScore });
   const annotationAnalytics = useAnnotationAnalytics();
   const fields = jiraTicket?.fields;
   const {
